Allow for different issuer specifiers in OpenAM 12 and 13
This patch fixes a failure to validate the ID token with OpenAM 13.

In OpenAM 12, the issuer path ends in at the base deployment path,
such as `/openam`.

In OpenAM 13, the issuer path ends in the base deployment path,
plus `/oauth2`, such as `/openam/oauth2`.

The workaround here is a little too lenient:
just check that the issuer starts with the correct path.